cybersecurity samples

+++ Solution5/DeleteMatche.php
@@ -2,8 +2,14 @@
 <?php
     require_once('../Connexion.inc.php'); //permet d'intégrer l'objet $conx à ce fichier
     try{
-        $conx->query("delete from matches where id=".$_GET['id']);
-        echo "Matche N°".$_GET['id']." bien supprimé";
+        if(isset($_SESSION['userName'])){
+            $conx->query("delete from matches where id=".$_GET['id']);
+            echo "Matche N°".$_GET['id']." bien supprimé";
+        }
+        else{
+            http_response_code(403);
+            die("Vous n'êtes pas autorisé à effectuer cette opération");
+        }
     }
     //Au cas ou une erreur a lieu pendant la suppression, elle sera affichée
     catch(Exception $e){
